+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 1.x branch
 ## 1.2 branch
+### 1.2.3
+* now you can set the `Users` class options (`className` and `foreignKey`)
+    using the `usersClassOptions` key of the `TokensTable` configuration;
+* small code fixes;
+* updated for `php-tools` 1.1.12.
+
 ### 1.2.2
 * `TokenTrait::find()`  method is now public;
 * updated for CakePHP 3.7.